% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Excel_Worksheet.R
\name{Build_Excel_Worksheet}
\alias{Build_Excel_Worksheet}
\title{Build Excel Worksheet}
\usage{
Build_Excel_Worksheet(
  SheetName,
  Data,
  WorkBook,
  CRDB_Columns = c(),
  Numeric_Columns = c(),
  Currency_Columns = c()
)
}
\arguments{
\item{DSN}{The Data Source Name (EDW_IBMi, EDW_External)}

\item{query}{The Query Text}

\item{linked_server_name}{The Name of the Linked Server (EDWtoIBMi)}
}
\value{
A Dataframe containing the Query Results
}
\description{
This function queries a Linked Server within SQL Server
}
\examples{
Query <- "Select * from table"
Data <- Query_Linked_Server(DSN = "EDWtoIBMi",
                            query = Query,
                            linked_server_name = "EDWtoIBMi")

}
