% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Location_Hierarchy.R
\name{Location_Hierarchy}
\alias{Location_Hierarchy}
\title{Location Hierarchy}
\usage{
Location_Hierarchy()
}
\value{
A Dataframe
}
\description{
Returns Lincare's Location Hierarchy
}
\examples{
Query <- "Select * from table"
Data <- Query_Linked_Server(DSN = "EDWtoIBMi",
                            query = Query,
                            linked_server_name = "EDWtoIBMi")

}
\references{
DBI
}
