% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Number_Format.R
\name{Number_Format}
\alias{Number_Format}
\title{Number Format}
\usage{
Number_Format(x, digits)
}
\arguments{
\item{x}{Numeric or field that be coerced to Numeric}

\item{digits}{How many digits should the field contain?}
}
\value{
A Vector containing the original number with leading zeros (nchar of output values should always be equal to the digits parameter)
}
\description{
This function converts a Number to a character string with leading zeros
}
\examples{
Number_Format(x = 8, digits = 2)
[1] "08"

}
