% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query_Linked_Server.R
\name{Query_Linked_Server}
\alias{Query_Linked_Server}
\title{Query Linked SQLServer}
\usage{
This can be used to Query the AS400/SYS38A or any other Linked Server
}
\arguments{
\item{DSN}{The Data Source Name (EDW_IBMi, EDW_External)
The Data Source Name (EDW_IBMi, EDW_External)}

\item{query}{The Query Text}

\item{linked_server_name}{The Name of the Linked Server (EDWtoIBMi)}
}
\value{
A Dataframe containing the Query Results
}
\description{
This function queries a Linked Server within SQL Server
}
\examples{
Query <- "Select * from table"
Data <- Query_Linked_Server(DSN = "EDWtoIBMi",
                            query = Query,
                            linked_server_name = "EDWtoIBMi")

}
\references{
RODBC
}
